requirements.txt
colorama
GitPython